Implement sentiment analysis (corenlp)
Substantiate the implementation of sentiment analysis using
Stanford's corenlp tools.

TODO: Increase performance -- currently, the sentiment classifier
is extremely slow. This is because PostClassifier launches the
java program for each post (instead of just once for all the posts).
The overhead of starting the program is on the order of two seconds,
and there are over 10,000 posts.

TODO: Make a better sentiment classifier. Either a) train the
Stanford model or b) build one using nltk classifiers.